test (mull-ui): update navigation test
The changes from Cristians branch should now have the updated tests

no ref
