+++ apps/mull-ui/src/app/components/event-card/event-card.spec.tsx
@@ -20,6 +20,6 @@ describe('EventCard', () => {
     const dom = render(<EventCard event={dummyEvent} />);
 
     const dateDiv = dom.getByTestId('event-card-datetime');
-    expect(dateDiv.textContent).toBe('1 Jan12:00AM');
+    expect(dateDiv.textContent).toBe('1 JAN12:00AM');
   });
 });
