+++ apps/mull-ui/src/app/components/navigation/bot-nav-bar/bot-nav-bar.spec.tsx
@@ -1,11 +1,10 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 
-import BottomNavBar from './header';
+import BottomNavBar from './bot-nav-bar';
 import { Router } from 'react-router-dom';
 
 import { createMemoryHistory } from 'history';
-import { ROUTES } from '../../../constants';
 
 import renderer from 'react-test-renderer';
 
@@ -20,21 +19,6 @@ describe('Header', () => {
     );
 
     expect(baseElement).toBeTruthy();
-  });
-
-  it('button associated with current should be active ', () => {
-    const testId = 'profile-mobile-navlink';
-
-    const history = createMemoryHistory();
-    const dom = render(
-      <Router history={history}>
-        <BottomNavBar />
-      </Router>
-    );
-
-    history.push(ROUTES.PROFILE);
-    const element = dom.getByTestId(testId);
-    expect(element.classList.contains('active')).toBeTruthy();
   });
 
   it('should match snapshot', () => {
